Fix missing finetune function.
The finetune function is no longer there in mcli apparently.
Instead, I found create_finetuning_run is there and seems to work fine. The arguments are slightly different, in that experiment_tracker definition takes a dict and is singular. Not sure what happened and why finetune isn't there any more, but this seemed to work for me for both tasks.